+++ NEWS.md
@@ -13,7 +13,8 @@
 - add stopFormatted()
 - improve implementation of hsAddMissingCols()
 - export noSuchElements()
-- improve extractSubstring() and subexpressionMatches()
+- improve extractSubstring() 
+- subexpressionMatches(): rename arg "regularExpression" to "pattern", improve
 - add nameElements()
 
 # [kwb.utils 0.13.0](https://github.com/KWB-R/kwb.utils/releases/tag/v0.13.0) <small>2022-06-04</small>